+++ src/main/java/codesquad/application/apihandler/UserHandler.java
@@ -1,7 +1,7 @@
-package codesquad.application.handler;
+package codesquad.application.apihandler;
 
 
-import codesquad.application.database.UserDatabase;
+import codesquad.application.datahandler.UserDatabase;
 import codesquad.application.domain.User;
 import codesquad.application.session.CookieExtractor;
 import codesquad.application.session.SessionManager;
